Small refactoring Quizzes, added icons quiz type
